// frontend/src/api/generated/model/employeeDocumentDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { EmployeeAttachmentDto } from './employeeAttachmentDto';

export interface EmployeeDocumentDto {
  id: string;
  documentType: string;
  /** @nullable */
  attachment?: EmployeeAttachmentDto | null;
  createdAt: string;
}